Recalculated Rules
Cleaned up some of the rules processing, re-tested to make sure each rule works independently.

Isolated one known issue: if a boid is approaching the right side of the screen at the middle line on a flat enough trajectory (close to 0 degrees) it will get trapped oscillating back and forth and leave the screen. To fix this, one should really track the direction the Boid _wants_ to be going and let it turn towards that if it's off the screen, rather than continuously re-calculating a new trajectory. I guess.

Signed-off-by: Seth Battis <seth@battis.net>